Final working version

+++ api/models/resource.py
@@ -1,8 +1,9 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String
 from api.dependencies.database import Base
 
 class Resource(Base):
     __tablename__ = "resources"
+
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(100), nullable=False)
-    quantity = Column(Float, nullable=False)
+    quantity = Column(Integer, nullable=False)
